pyproject.toml
[project]
name = "contabilita-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "python-codicefiscale>=0.10.4",
]
